+++ scripts/carride/main.gd
@@ -1,15 +1,18 @@
 extends Node2D
 
-var ai_vehicle_instance := preload("res://nodes/PandaAI.tscn")
-var milk_instance := preload("res://nodes/Milk.tscn")
-
-@onready var SpeedometerGauge: Sprite2D = $Camera2D/Speedometer/SpeedometerGauge
-@onready var GUI: Control = $GUI
-@onready var MilkSpawner: Timer = $Timer
-
 signal vehicle_reached_oob
 
 @export var speed_increment: float = 0.2
+
+@export var spawn_time_limit: int = 20:
+	get:
+		return spawn_time_limit
+@export var speed: float = 0:
+	get:
+		return speed
+
+var ai_vehicle_instance := preload("res://nodes/PandaAI.tscn")
+var milk_instance := preload("res://nodes/Milk.tscn")
 var stop_processing := false
 var spawn_time: int = 0
 var gauge_speed: float = -36
@@ -18,12 +21,8 @@ var milks: int = 0:
 	get:
 		return milks
 	set(value):
-		milks = value 
-		GUI.update_points()
-
-@export var spawn_time_limit: int = 20:
-	get:
-		return spawn_time_limit
+		milks = value
+		gui.update_points()
 
 var spawn_rnd_min: int = 0:
 	get:
@@ -32,27 +31,22 @@ var spawn_rnd_min: int = 0:
 var spawn_rnd_max: int = 1:
 	get:
 		return spawn_rnd_max + ((speed - 300) / 300)
-		
+
 var start_time: int:
 	get:
 		return start_time
+
+var time: int:
+	get:
+		return time
 
 var final_score: int:
 	get:
 		return final_score
 
-@export var speed: float = 0:
-	get:
-		return speed
-
-func game_over():
-	final_score = (milks * 100000) + roundi((Time.get_ticks_msec() - start_time) * 0.01)
-	GUI.game_over()
-	$PauseLayer.process_mode = Node.PROCESS_MODE_DISABLED
-	stop_processing = true
-	if not SettingsBus.cheats:
-		var leaderboard_name = "dev" if OS.is_debug_build() else "main"
-		SilentWolf.Scores.save_score(SettingsBus.playername, final_score, leaderboard_name)
+@onready var speedometer_gauge: Sprite2D = $Camera2D/Speedometer/SpeedometerGauge
+@onready var gui: Control = $GUI
+@onready var milk_spawner: Timer = $Timer
 
 
 func _ready():
@@ -63,20 +57,22 @@ func _ready():
 
 
 func _process(delta):
+	time = Time.get_ticks_msec() - start_time
 	_change_scale_factor() #FIXME: Find better solution
 	if not stop_processing:
 		if speed < 1560:
 			speed += delta * speed_increment
 			gauge_speed = (speed / 10) - 66
-			SpeedometerGauge.rotation_degrees = gauge_speed
+			speedometer_gauge.rotation_degrees = gauge_speed
 	else:
 		if speed > -300:
 			speed -= pow(delta, 2) * 10000
 		else:
 			speed = -300
+	_achievement_check()
 
 
-func _physics_process(delta):
+func _physics_process(_delta):
 	spawn_time += randi_range(spawn_rnd_min,spawn_rnd_max)
 	if spawn_time > spawn_time_limit:
 		var loc_rnd = randi_range(0,4)
@@ -88,6 +84,16 @@ func _physics_process(delta):
 			car.position.x = marker.get_parent().position.x
 			add_child(car)
 			car.owner = self
+
+
+func game_over():
+	final_score = (milks * 100000) + roundi((Time.get_ticks_msec() - start_time) * 0.01)
+	gui.game_over()
+	$PauseLayer.process_mode = Node.PROCESS_MODE_DISABLED
+	stop_processing = true
+	if not SettingsBus.cheats:
+		var leaderboard_name = "dev" if OS.is_debug_build() else "main"
+		SilentWolf.Scores.save_score(SettingsBus.playername, final_score, leaderboard_name)
 
 
 func _change_scale_factor():
@@ -110,3 +116,11 @@ func _respawn_vehicle(): # Is this needed anymore?
 	y += 100 * randi_range(0,2)
 	$"PandaBlue".position = Vector2(1450, y)
 
+
+func _achievement_check():
+	if speed > 800:
+		AchievementSystem.call_achievement("speed_110")
+	if speed > 1200:
+		AchievementSystem.call_achievement("speed_150")
+	if time > 60000 and milks == 0:
+		AchievementSystem.call_achievement("lactose_intolerant")