+++ scripts/main_menu.gd
@@ -1,12 +1,6 @@
 extends Control
 
-signal DEBUG_DEV_BUTTONS
-
-
-@onready var options_layer = $OptionsLayer
-@onready var http_request = $HTTPRequest
-@onready var http_request_changelog = $HTTPRequest2
-@onready var update_button = $MainMenuLayer/UpdateButton
+signal signal_debug_dev_buttons
 
 @export var skeletonek_logo: Control
 @export var godot_logo: Control
@@ -16,38 +10,49 @@ var version: String
 var download_url: String
 var changelog: String
 
+@onready var options_layer = $OptionsLayer
+@onready var http_request = $HTTPRequest
+@onready var http_request_changelog = $HTTPRequest2
+@onready var update_button = $MainMenuLayer/UpdateButton
+
 
 func _ready():
 	if GlobalMusic.stream.resource_path == "res://audio/music/the-great-rescue.ogg":
-		var musicTick = GlobalMusic.get_playback_position()
+		var music_tick = GlobalMusic.get_playback_position()
 		GlobalMusic.stream = load("res://audio/music/the-greatest-rescue.ogg")
 		GlobalMusic.play()
-		GlobalMusic.seek(musicTick)
+		GlobalMusic.seek(music_tick)
 
 	skeletonek_logo.gui_input.connect(_on_skeletonek_logo_click)
 	godot_logo.gui_input.connect(_on_godot_logo_click)
-	
+
 #	GlobalMusic.finished.connect(_on_global_music_finished)
 #	if not GlobalMusic.is_playing():
 #		_on_global_music_finished()
 	$MainMenuLayer/VBoxContainer/ContinueButton.grab_focus()
-	
-	DEBUG_DEV_BUTTONS.connect(_enable_dev_buttons)
+
+	signal_debug_dev_buttons.connect(_enable_dev_buttons)
 	# Developer console is inaccessible on Android, unless you have plugged in a keyboard
 	if OS.get_name() == "Android" and OS.is_debug_build():
 		$MainMenuLayer/DevButtons.visible = true
-	
-	var device_info_label = $CreditsLayer/PanelContainer/MarginContainer/VBoxContainer/ScrollContainer/HBoxContainer/VBoxContainer2/DeviceInfo
+
+	var device_info_label = $CreditsLayer/PanelContainer/MarginContainer/\
+		VBoxContainer/ScrollContainer/HBoxContainer/VBoxContainer2/DeviceInfo
 	device_info_label.text = DebugInfo.debug_info() + "\n\n"
-	var game_info_label = $CreditsLayer/PanelContainer/MarginContainer/VBoxContainer/ScrollContainer/HBoxContainer/VBoxContainer2/GameInfo
-	game_info_label.text = "Wersja gry: " + ProjectSettings.get_setting("application/config/version") + "\n\n"
+	var game_info_label = $CreditsLayer/PanelContainer/MarginContainer/\
+		VBoxContainer/ScrollContainer/HBoxContainer/VBoxContainer2/GameInfo
+	game_info_label.text = "Wersja gry: " + \
+		ProjectSettings.get_setting("application/config/version") + "\n\n"
 
 	# Check for update
 	http_request.request_completed.connect(_validate_update)
 	http_request_changelog.request_completed.connect(_download_changelog)
 	_check_update()
 
-	get_parent().animation.button.pressed.connect(_on_easter_egg_button_pressed)
+	if not get_parent() is Window:
+		get_parent().animation.button.pressed.connect(_on_easter_egg_button_pressed)
+	else:
+		$MenuAnim.button.pressed.connect(_on_easter_egg_button_pressed)
 
 
 func _on_global_music_finished():
@@ -55,6 +60,7 @@ func _on_global_music_finished():
 	print("Unix Timestamp % 2 = ", date)
 	match(date):
 		0:
+			# gdlint:ignore=duplicated-load
 			GlobalMusic.stream = load("res://audio/music/the-greatest-rescue.ogg")
 		1:
 			GlobalMusic.stream = load("res://audio/music/tech-junkie.ogg")
@@ -70,14 +76,13 @@ func _check_update():
 	print("Checking if update is available...")
 
 
-func _validate_update(result, response_code, headers, body):
+func _validate_update(result, _response_code, _headers, body):
 	if result != 0:
 		push_warning("Couldn't check for update! Do we have an internet?")
 		update_button.text = "Brak połączenia"
 		update_button.disabled = true
 		update_button.show()
 		get_tree().create_timer(10).timeout.connect(_check_update)
-		pass # There is a connection error
 	else:
 		print("Correctly checked if update is available")
 		var json = JSON.parse_string(body.get_string_from_utf8())
@@ -85,18 +90,18 @@ func _validate_update(result, response_code, headers, body):
 			print("Update found!")
 			download_url = json["downloadURL"]
 			version = json["buildCode"]
-			
+
 			var changelog_url = json["changelogURL"]
 			push_warning(changelog_url)
 			http_request_changelog.request(changelog_url)
-			
+
 			$PopupChangelog.update_version_data(version, download_url)
 			update_button.text = "Dostępna aktualizacja!"
 			update_button.disabled = false
 			update_button.show()
 
 
-func _download_changelog(result, response_code, headers, body):
+func _download_changelog(_result, _response_code, _headers, body):
 	$PopupChangelog.update_changelog_data(body.get_string_from_utf8())
 
 
@@ -130,6 +135,12 @@ func _on_credits_button_pressed():
 	$CreditsLayer/PanelContainer/MarginContainer/VBoxContainer/BackButton.grab_focus()
 
 
+func _on_achievements_button_pressed():
+	_hide_all_layers()
+	$AchievementsLayer.show()
+	$AchievementsLayer/PanelContainer/VBoxContainer/MarginContainer2/BackButton.grab_focus()
+
+
 func _on_hi_score_button_pressed():
 	_hide_all_layers()
 	$LeaderboardLayer.show()
@@ -151,6 +162,7 @@ func _hide_all_layers():
 	$MainMenuLayer.hide()
 	$OptionsLayer.hide()
 	$LeaderboardLayer.hide()
+	$AchievementsLayer.hide()
 	$CreditsLayer.hide()