+++ scripts/ui/achievements_viewer.gd
@@ -16,7 +16,8 @@ var achievement_container = preload("res://scenes/ui/achievement/achievement_con
 var achievement_animation = preload("res://scenes/ui/achievement/animation_glow.tscn")
 var locked_achv_icon = preload("res://images/icons/slb2icon.png")
 var desaturate_material = preload("res://shaders/desaturation_material.tres")
-@onready var achv_panel = $PanelContainer/VBoxContainer/MarginContainer/ScrollContainer/HFlowContainer
+@onready var achv_panel = $PanelContainer/VBoxContainer/MarginContainer/\
+	ScrollContainer/HFlowContainer
 @onready var achv_details = $AchvDetailsView
 
 func _ready():
@@ -79,7 +80,13 @@ func _generate_hidden_achievements():
 					)
 
 
-func _generate_achievement_containers(id: String, title: String, icon: CompressedTexture2D, glow: bool, desaturate = false):
+func _generate_achievement_containers(
+			id: String,
+			title: String,
+			icon: CompressedTexture2D,
+			glow: bool,
+			desaturate = false
+	):
 	var instance = achievement_container.instantiate()
 	instance.get_node("AchievementIcon").texture = icon
 	if desaturate:
@@ -88,7 +95,7 @@ func _generate_achievement_containers(id: String, title: String, icon: Compresse
 	if glow:
 		var anim_instance = achievement_animation.instantiate()
 		instance.get_node("AchievementIcon").add_child(anim_instance)
-	
+
 	instance.achievement_index = id
 	instance.gui_input.connect(_load_achv_details.bind(instance))
 	achv_panel.add_child(instance)
@@ -99,11 +106,11 @@ func _load_achv_details(event, sender):
 		if event.button_index == MOUSE_BUTTON_LEFT and event.pressed:
 			var achv_index = sender.achievement_index
 			var achv_data = AchievementSystem.data[achv_index]
-			
+
 			if sender.achievement_anim != null:
 				AchievementSystem.recently_completed[achv_index] = false
 				sender.achievement_anim.queue_free()
-				
+
 			if achv_data[ACHV_HIDDEN] and not (
 			AchievementSystem.completed[achv_index][ACHV_COMPLETE]):
 				achv_details.title.text = "Ukryte osiągnięcie"
@@ -115,8 +122,8 @@ func _load_achv_details(event, sender):
 				achv_details.desc.text = achv_data[ACHV_DESC]
 				achv_details.icon.texture = achv_data[ACHV_ICON]
 				achv_details.date.text = "Nie odblokowano" if not (
-				AchievementSystem.completed[achv_index][ACHV_COMPLETE]) else (
-				AchievementSystem.completed[achv_index][ACHV_DATE]
+						AchievementSystem.completed[achv_index][ACHV_COMPLETE]) else (
+						AchievementSystem.completed[achv_index][ACHV_DATE]
 				)
 			achv_details.visible = true
 			print("Clicked ", sender.get_node("AchievementTitle").text)
@@ -125,6 +132,6 @@ func _load_achv_details(event, sender):
 func _process(_delta):
 	pass
 
-
 func _on_back_button_pressed():
+	# gdlint:ignore=private-method-call
 	$"../"._on_back_button_pressed()