+++ scripts/ui/nodes/SwitchButton.gd
@@ -6,7 +6,8 @@ func _ready():
 func _toggled(button_pressed):
 	_change_text(button_pressed)
 
-## This function does the same as [method BaseButton.set_pressed_no_signal] but also changes text according to [code]pressed[/code] state
+## This function does the same as [method BaseButton.set_pressed_no_signal]
+## but also changes text according to [code]pressed[/code] state
 func set_pressed_no_signal_custom(pressed: bool):
 	_change_text(pressed)
 	set_pressed_no_signal(pressed)